Почистил проект от дубликатов
