+++ src/main/java/faang/school/projectservice/mapper/moment/MomentMapper.java
@@ -1,6 +1,7 @@
 package faang.school.projectservice.mapper.moment;
 
 import faang.school.projectservice.dto.moment.MomentDto;
+import faang.school.projectservice.mapper.ProjectMapper;
 import faang.school.projectservice.model.Moment;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
