+++ src/main/java/faang/school/projectservice/service/meet/MeetService.java
@@ -15,8 +15,8 @@ import faang.school.projectservice.repository.MeetRepository;
 import faang.school.projectservice.repository.ProjectRepository;
 import faang.school.projectservice.service.meet.filter.MeetFilter;
 import faang.school.projectservice.service.team.TeamService;
+import faang.school.projectservice.validator.ProjectValidator;
 import faang.school.projectservice.validator.meet.MeetValidator;
-import faang.school.projectservice.validator.project.ProjectValidator;
 import faang.school.projectservice.validator.team.TeamValidator;
 import faang.school.projectservice.validator.user.UserValidator;
 import lombok.Builder;
