+++ src/main/resources/application.yaml
@@ -26,7 +26,7 @@ spring:
           name: calculations_channel
 
 server:
-  port: 8082
+  port: 8081
 
 logging:
   level:
